old-tests/submission/Infraware/Forms/contents/Forms/button_name.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>button_name</h3>
	</p>
	
	<hr>

	<div id="log"></div> 

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  name="input_form">
	<p><button id='button_id' name='name'>button</button></p>
	</form>

	<script>
		
		var button = document.getElementById("button_id");
		name = button.name;
		button.name = "button_name";

		if (typeof(button.name) == "string") {
			if (name == 'name' && button.name == "button_name") {
				test(function() {
					assert_true(true, "name attribute is correct.");
				});
			} else {
				test(function() {
					assert_unreached("name attribute is not correct.");
				});
			}
		} else {
			test(function() {
				assert_unreached("name attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
